Even more refactoring
