+++ package.json
@@ -33,6 +33,7 @@
     "unescape": "^1.0.1"
   },
   "devDependencies": {
+    "@enonic-types/core": "^7.14.0-B2",
     "@enonic/eslint-config": "^1.1.0",
     "@types/negotiator": "^0.6.3",
     "@types/node": "^20.8.9",
