+++ src/guillotine/getParentRegion.ts
@@ -6,7 +6,7 @@ import type {
 } from '../types';
 
 
-import {XP_COMPONENT_TYPE} from '../utils';
+import {XP_COMPONENT_TYPE} from '../constants';
 import {prefixLayoutPath} from './prefixLayoutPath';
