+++ src/guillotine/parseComponentPath.ts
@@ -4,7 +4,7 @@ import type {PathFragment} from '../types';
 import {
     FRAGMENT_CONTENTTYPE_NAME,
     FRAGMENT_DEFAULT_REGION_NAME,
-} from '../utils';
+} from '../constants';
 
 
 export function parseComponentPath(contentType: string, path: string): PathFragment[] {
