+++ src/utils/fixDoubleSlashes.ts
@@ -1,0 +1,3 @@
+export function fixDoubleSlashes(str: string) {
+    return str.replace(/(^|[^:/])\/{2,}/g, '$1/');
+}
