+++ src/utils/getProjectLocales.ts
@@ -1,0 +1,6 @@
+import {getProjectLocaleConfigs} from './getProjectLocaleConfigs';
+
+
+export function getProjectLocales(): string[] {
+    return Object.keys(getProjectLocaleConfigs());
+}
