+++ test/constants.ts
@@ -1,2 +1,4 @@
+export const ENONIC_API = 'http://localhost:8080/site';
 export const ENONIC_APP_NAME = 'com.enonic.app.enonic';
 export const ENONIC_APP_NAME_UNDERSCORED = ENONIC_APP_NAME.replace(/\./g, '_');
+export const ENONIC_PROJECTS = 'en:project/site,no:prosjekt/nettsted';
