+++ test/guillotine/fetchGuillotine.test.ts
@@ -104,7 +104,7 @@ describe('guillotine', () => {
 
     describe('fetchGuillotine', () => {
         it('fetches a response from guillotine', () => {
-            import('../../src/guillotine/fetchGuillotine').then((moduleName) => {
+            import('../../src').then((moduleName) => {
                 moduleName.fetchGuillotine(...FETCH_GUILLOTINE_PARAMS_VALID).then((result) => {
                     expect(result).toEqual({
                         guillotine: {
@@ -126,7 +126,7 @@ describe('guillotine', () => {
                 FETCH_GUILLOTINE_PARAMS_VALID[2],
                 FETCH_GUILLOTINE_PARAMS_VALID[3],
             ]
-            import('../../src/guillotine/fetchGuillotine').then((moduleName) => {
+            import('../../src').then((moduleName) => {
                 moduleName.fetchGuillotine(contentApiUrl, body, projectConfig, headers).then((result) => {
                     expect(result).toEqual({
                         guillotine: {
@@ -149,7 +149,7 @@ describe('guillotine', () => {
                     status: 200
                 } as Response);
             });
-            import('../../src/guillotine/fetchGuillotine').then((moduleName) => {
+            import('../../src').then((moduleName) => {
                 moduleName.fetchGuillotine(...FETCH_GUILLOTINE_PARAMS_VALID).then((result) => {
                     expect(result).toEqual({
                         error: {
@@ -173,7 +173,7 @@ describe('guillotine', () => {
                 FETCH_GUILLOTINE_PARAMS_VALID[2],
                 FETCH_GUILLOTINE_PARAMS_VALID[3],
             ]
-            import('../../src/guillotine/fetchGuillotine').then((moduleName) => {
+            import('../../src').then((moduleName) => {
                 moduleName.fetchGuillotine(contentApiUrl, body, projectConfig, headers).then((result) => {
                     expect(result).toEqual({
                         guillotine: {
@@ -196,7 +196,7 @@ describe('guillotine', () => {
                 FETCH_GUILLOTINE_PARAMS_VALID[2],
                 FETCH_GUILLOTINE_PARAMS_VALID[3]
             ]
-            import('../../src/guillotine/fetchGuillotine').then((moduleName) => {
+            import('../../src').then((moduleName) => {
                 moduleName.fetchGuillotine(contentApiUrl, body, projectConfig, headers).then((result) => {
                     expect(result).toEqual({
                         error: {
@@ -218,7 +218,7 @@ describe('guillotine', () => {
                 FETCH_GUILLOTINE_PARAMS_VALID[2],
                 FETCH_GUILLOTINE_PARAMS_VALID[3],
             ]
-            import('../../src/guillotine/fetchGuillotine').then((moduleName) => {
+            import('../../src').then((moduleName) => {
                 moduleName.fetchGuillotine(contentApiUrl, body, projectConfig, headers).then((result) => {
                     expect(result).toEqual({
                         error: {
@@ -242,7 +242,7 @@ describe('guillotine', () => {
                     status: 200
                 } as Response);
             });
-            import('../../src/guillotine/fetchGuillotine').then((moduleName) => {
+            import('../../src').then((moduleName) => {
                 moduleName.fetchGuillotine(...FETCH_GUILLOTINE_PARAMS_VALID).then((result) => {
                     expect(result).toEqual({
                         error: {
@@ -258,7 +258,7 @@ describe('guillotine', () => {
             jest.spyOn(globalThis, 'fetch').mockImplementation(() => {
                 throw new Error('fetch error');
             });
-            import('../../src/guillotine/fetchGuillotine').then((moduleName) => {
+            import('../../src').then((moduleName) => {
                 expect(moduleName.fetchGuillotine(...FETCH_GUILLOTINE_PARAMS_VALID)).resolves.toEqual({
                     error: {
                         code: 'API',
@@ -277,7 +277,7 @@ describe('guillotine', () => {
                     status: 500
                 } as Response);
             });
-            import('../../src/guillotine/fetchGuillotine').then((moduleName) => {
+            import('../../src').then((moduleName) => {
                 expect(moduleName.fetchGuillotine(...FETCH_GUILLOTINE_PARAMS_VALID)).resolves.toEqual({
                     error: {
                         code: 500,
@@ -295,7 +295,7 @@ describe('guillotine', () => {
                     status: 200
                 } as Response);
             });
-            import('../../src/guillotine/fetchGuillotine').then((moduleName) => {
+            import('../../src').then((moduleName) => {
                 expect(moduleName.fetchGuillotine(...FETCH_GUILLOTINE_PARAMS_VALID)).resolves.toEqual({
                     error: {
                         code: 500,
@@ -314,7 +314,7 @@ describe('guillotine', () => {
                     status: 200
                 } as Response);
             });
-            import('../../src/guillotine/fetchGuillotine').then((moduleName) => {
+            import('../../src').then((moduleName) => {
                 expect(moduleName.fetchGuillotine(...FETCH_GUILLOTINE_PARAMS_VALID)).resolves.toEqual({
                     error: {
                         code: 500,
@@ -332,7 +332,7 @@ describe('guillotine', () => {
             jest.doMock<typeof import('../../src/guillotine/fetchFromApi')>(
                 '../../src/guillotine/fetchFromApi', () => ({fetchFromApi})
             );
-            import('../../src/guillotine/fetchGuillotine').then((moduleName) => {
+            import('../../src').then((moduleName) => {
                 expect(moduleName.fetchGuillotine(...FETCH_GUILLOTINE_PARAMS_VALID)).resolves.toEqual({
                     error: {
                         code: 'Client-side error',
