+++ test/tsconfig.json
@@ -1,6 +1,6 @@
 {
     "compilerOptions": {
         "esModuleInterop": true,
-        "jsx": "preserve"
+        "jsx": "react"
     }
 }
